# public/docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Asparecargas Endpoints",
        "description": "Here are the endpoints of Asparecargas",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "https://testing.asparecargas.net"
        }
    ],
    "paths": {
        "/api/v1/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get enabled products",
                "operationId": "52f07b5c504cf5dead737602311e23a8",
                "responses": {
                    "200": {
                        "description": "Get enabled products",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "type": {
                                                        "type": "string",
                                                        "example": "product"
                                                    },
                                                    "attributes": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "example": 1
                                                            },
                                                            "product_name": {
                                                                "type": "string",
                                                                "example": "BANCOLOMBIA"
                                                            },
                                                            "product_description": {
                                                                "type": "string",
                                                                "example": "<p>DEPOSITO BANCOLOMBIA</p>"
                                                            },
                                                            "product_logo": {
                                                                "type": "string",
                                                                "example": "https://example.com/storage/products/1.png"
                                                            },
                                                            "product_commission": {
                                                                "type": "integer",
                                                                "example": 800
                                                            },
                                                            "product_type": {
                                                                "type": "string",
                                                                "example": "Deposit"
                                                            },
                                                            "client_name": {
                                                                "type": "integer",
                                                                "example": 0
                                                            },
                                                            "client_document": {
                                                                "type": "integer",
                                                                "example": 0
                                                            },
                                                            "phone_number": {
                                                                "type": "integer",
                                                                "example": 0
                                                            },
                                                            "email": {
                                                                "type": "integer",
                                                                "example": 0
                                                            },
                                                            "account_type": {
                                                                "type": "integer",
                                                                "example": 1
                                                            },
                                                            "account_number": {
                                                                "type": "integer",
                                                                "example": 1
                                                            },
                                                            "code": {
                                                                "type": "integer",
                                                                "example": 0
                                                            },
                                                            "extra": {
                                                                "type": "integer",
                                                                "example": 0
                                                            },
                                                            "are_default_fields": {
                                                                "type": "integer",
                                                                "example": 1
                                                            },
                                                            "field_names": {
                                                                "type": "string",
                                                                "example": "Número de seguro,Número de NIT,Número de registro civil"
                                                            },
                                                            "min_amount": {
                                                                "type": "integer",
                                                                "example": 5000
                                                            },
                                                            "max_amount": {
                                                                "type": "integer",
                                                                "example": 300000
                                                            },
                                                            "category": {
                                                                "type": "string",
                                                                "example": "bancos"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/v1/transaction/create": {
            "post": {
                "tags": [
                    "Transaction"
                ],
                "summary": "Create Transaction",
                "operationId": "1638ec919252ee286da613c58cdc91b6",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "product_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "account_number": {
                                        "type": "string",
                                        "example": "1234567890"
                                    },
                                    "amount": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 20000.5
                                    },
                                    "type": {
                                        "$ref": "#/components/schemas/TransactionType"
                                    },
                                    "detail": {
                                        "type": "string",
                                        "example": "Tipo de cuenta: Ahorros,Nit: 332435,"
                                    },
                                    "date": {
                                        "type": "string",
                                        "example": "2024-09-10"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success Transaction Creation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "transaction_id": {
                                            "type": "integer",
                                            "example": "3111"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Transacción asignada a proveedor"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "204": {
                        "description": "There are no suppliers",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "transaction_id": {
                                            "type": "integer",
                                            "example": "3111"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "No hay proveedores disponibles"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "429": {
                        "description": "Transaction limit per period exceeded",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "transaction_id": {
                                            "type": "integer",
                                            "example": "3111"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Esta cuenta superó el límite de transacciones por periodo, podrá realizar transacciones con la misma a partir de las 11:06 PM"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/transaction/detail/{id}": {
            "get": {
                "tags": [
                    "Transaction"
                ],
                "summary": "Get transaction detail",
                "operationId": "466a752f509d1a679e0cc3e3727460ee",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the transaction to retrieve",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get transaction detail",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "type": {
                                                    "type": "string",
                                                    "example": "transaction_detail"
                                                },
                                                "attributes": {
                                                    "properties": {
                                                        "transaction_id": {
                                                            "type": "integer",
                                                            "example": 3817
                                                        },
                                                        "transaction_status": {
                                                            "type": "string",
                                                            "example": "cancelled"
                                                        },
                                                        "account_number": {
                                                            "type": "string",
                                                            "example": "1234567810"
                                                        },
                                                        "transaction_type": {
                                                            "type": "string",
                                                            "example": "Deposit"
                                                        },
                                                        "amount": {
                                                            "type": "number",
                                                            "format": "float",
                                                            "example": 20000.5
                                                        },
                                                        "transaction_detail": {
                                                            "type": "string",
                                                            "example": "Tipo de cuenta: Ahorros,Nit: 332435,"
                                                        },
                                                        "product_id": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "product_name": {
                                                            "type": "string",
                                                            "example": "BANCOLOMBIA"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Error: Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/user/login": {
            "post": {
                "tags": [
                    "Login"
                ],
                "summary": "User Login",
                "operationId": "f8cb634ff438083cd231d25a056e3497",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success Login",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Success Login"
                                        },
                                        "access_token": {
                                            "type": "string",
                                            "example": "eyJ0eXAiOiJKV1QiLCJhbGciOiJSUzI1NiJ9.eyJhdWQiOiIxIiwianRpIjoiNWM5NzZhMWJkYjNhMjdiNjBjZDEyNWU2OTQ5ZWY5YjQ3MWE1ZTA4MzQyODJlNDAwNmU4MzJkODY5ODYyMTg0M2FhMzkzMDMwYjI1YzM1MGYiLCJpYXQiOjE3MTY2OTMyMzQuNzA2NjksIm5iZiI6MTcxNjY5MzIzNC43MDY2OTUsImV4cCI6MTc0ODIyOTIzNC40MjkzODcsInN1YiI6IjI3Iiwic2NvcGVzIjpbXX0.5XcEdol8IOx7Z8ot3jnK9BsmIRlTo5mDUSTcl3dOOlTY40kWmi26OULLjwXLDBWVW0SmwXxYM8c6zhpXo0I-ahEAmUMrIXtzYQGZIV4fsJeFjnVyBUJocTSydfp96dmYPNxgqbYHYIR-dZaBSrXKFU1yJ_R3ozz_1K4UlGv2i5eDHOSmqMS_vFTEQWUeTYuDc9gppmqTOn9IV3CXQRGi_8cTlHhdSGlb2Fx8rj2ccUh5Lenrc59A79euPEl0Gh09xJZLvaIFd0ZvwCQnmA-ZVhsv5CRk0ReC4QGBqOPpOmPc_z_yK_F5gpiUMGEWzZy33tn1JVe92sfdAhpGyYulr4Kpxa8cpgvOSQ45HyE0bwpoSL0008S_nEQiMTdvnNr-lcdyMRpYM__nanXiOpQFl3U3N-1R5oiOb83-bh0U5elrdqnhmEECiluFN1T2Mkd97C4IeHmA4lpHI8dZGVyYeou34JKG30Ht2qeU8V84Gtn6DX6JZWzQjgk9Ryn4PlecA-rIuQdLzeSWbQpDTuInuz0s51srnIY4s0FIoTCFNVOdK10Tkd9eRHk-UDJvaufk8FapNqMWIn16PjLbQ9hnRMdkKq9JwmIvOkq7Ory96h0JuyAQ0SLxP6GdjNhp8rtEiqbyNtLGUTNrfPmhw18HP-3rQ4mlKPdrym1U2gdjqAM"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Invalid Credentials"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden Access",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Forbidden Access"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "TransactionType": {
                "type": "string",
                "enum": [
                    "Deposit",
                    "Withdrawal"
                ]
            }
        },
        "securitySchemes": {
            "passport": {
                "type": "apiKey",
                "description": "Ingrese el token en el formato (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "tags": [
        {
            "name": "Products",
            "description": "Products"
        },
        {
            "name": "Transaction",
            "description": "Transaction"
        },
        {
            "name": "Login",
            "description": "Login"
        }
    ],
    "security": [
        {
            "passport": []
        }
    ]
}